DEAPP fixes copied from Postgres test server

+++ ddl/postgres/deapp/de_pathway_gene.sql
@@ -3,7 +3,7 @@
 --
 CREATE TABLE de_pathway_gene (
     id bigint NOT NULL,
-    pathway_id bigint,
+    pathway_id numeric,
     gene_symbol character varying(200),
     gene_id character varying(200)
 );
